requirements.txt
pytest
playwright
pytest-playwright
python-dotenv
requests
allure-pytest